Refs #106. fixing tests
