+++ packages/mcni/tests/mcni/pyre_support/parallel_app.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# Jiao Lin <jiao.lin@gmail.com
+# Jiao Lin <jiao.lin@gmail.com>
 #
 
 import journal
