+++ packages/mcni/tests/mcni/pyre_support/parallel_simulation_TestCase.pml
@@ -1,16 +1,4 @@
 <?xml version="1.0"?>
-<!--
-! ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-!
-!                                   Jiao Lin
-!                      California Institute of Technology
-!                        (C) 2008  All Rights Reserved
-!
-! {LicenseText}
-!
-! ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
--->
-
 
 <!DOCTYPE inventory>
 
@@ -21,19 +9,8 @@
     <property name="buffer_size">1</property>
     <property name="output-dir">parallel_simulation_test_out</property>
     <property name="overwrite-datafiles">1</property>
-
-    <component name="mpirun">
-       <property name="nodes">2</property>
-    </component>
-
   </component>
 
 </inventory>
-
-
-<!-- version-->
-<!-- $Id$-->
-
-<!-- Generated automatically by XMLMill on Fri Jul  4 10:11:52 2008-->
 
 <!-- End of file -->
